Remove www mapping and add S3 bucket for redirects
This will let us use the S3 bucket (and CloudFront in front of it, with
the proper certs) to redirect to the main web domain. This will mean we
never have to worry about issues with CORS from multiple domains, or
worry about defining/creating individual mappings for each supported
domain.