Dynamically add static routes to SAM template
Kinda a cheat, but I don't want to be maintaining this list in 2 places.
While I would have rather just loaded the YAML and modified it, the
loader stumbles over the CFN functions/references and I couldn't find an
acceptible way to get it to work and output valid data.