Some test code for generating XSRF tokens
Just some playing around to see what I can get going here, using local
DynamoDB. The idea here is we generate the XSRF using a random UUID,
and send that to the client, while storing a copy in the table that's
further mixed with the form name and the IP of the user that requested
it.

This avoids us storing that information along with the tokens, while
still ensuring they match.